Convert from float64 to float32

+++ chemprop/cli/train.py
@@ -752,8 +752,12 @@ def train_model(args, train_loader, val_loader, test_loader, output_dir, scaler,
 
         if test_loader is not None:
             if args.task_type == "regression":
-                model.predictor.register_buffer("loc", torch.tensor(scaler.mean_).view(1, -1))
-                model.predictor.register_buffer("scale", torch.tensor(scaler.scale_).view(1, -1))
+                model.predictor.register_buffer(
+                    "loc", torch.tensor(scaler.mean_).view(1, -1).float()
+                )
+                model.predictor.register_buffer(
+                    "scale", torch.tensor(scaler.scale_).view(1, -1).float()
+                )
             results = trainer.test(model, test_loader)[0]
             logger.info(f"Test results: {results}")
